+++ src/common/Channel.cpp
@@ -71,15 +71,24 @@ bool Channel::hasMessages() const
     return !this->messages_.empty();
 }
 
-LimitedQueueSnapshot<MessagePtr> Channel::getMessageSnapshot() const
+std::vector<MessagePtr> Channel::getMessageSnapshot() const
 {
     return this->messages_.getSnapshot();
 }
 
-LimitedQueueSnapshot<MessagePtr> Channel::getMessageSnapshot(
-    size_t nItems) const
+std::vector<MessagePtr> Channel::getMessageSnapshot(size_t nItems) const
 {
     return this->messages_.lastN(nItems);
+}
+
+MessagePtr Channel::getLastMessage() const
+{
+    auto last = this->messages_.last();
+    if (last)
+    {
+        return *std::move(last);
+    }
+    return nullptr;
 }
 
 void Channel::addMessage(MessagePtr message, MessageContext context,
@@ -147,19 +156,15 @@ void Channel::addOrReplaceClearChat(MessagePtr message, const QDateTime &now)
 
 void Channel::disableAllMessages()
 {
-    LimitedQueueSnapshot<MessagePtr> snapshot = this->getMessageSnapshot();
-    int snapshotLength = snapshot.size();
-    for (int i = 0; i < snapshotLength; i++)
+    for (const auto &message : this->getMessageSnapshot())
     {
-        const auto &message = snapshot[i];
         if (message->flags.hasAny({MessageFlag::System, MessageFlag::Timeout,
                                    MessageFlag::Whisper}))
         {
             continue;
         }
 
-        // FOURTF: disabled for now
-        const_cast<Message *>(message.get())->flags.set(MessageFlag::Disabled);
+        message->flags.set(MessageFlag::Disabled);
     }
 }